+++ src/components/MyProfile.jsx
@@ -4,7 +4,7 @@ import '../styles/Profile.css';
 const MyProfile = () => (
   <>
     <h1>My Profile</h1>
-    <h2>Under Construction</h2>
+    <h2>Page Under Construction</h2>
   </>
 );
