+++ src/components/Navigation.jsx
@@ -4,7 +4,7 @@ import Container from 'react-bootstrap/Container';
 import Navbar from 'react-bootstrap/Navbar';
 import Nav from 'react-bootstrap/Nav';
 import { NavLink } from 'react-router-dom';
-import logo from '../assets/logo.png';
+import planet from '../assets/planet.png';
 import NavStyle from '../styles/Navbar.module.css';
 
 const Navbarmenu = () => (
@@ -12,7 +12,7 @@ const Navbarmenu = () => (
     <Navbar bg="light" variant="light">
       <Container className={NavStyle.navContainer}>
         <Navbar.Brand className={NavStyle.navTitle}>
-          <img alt="Logo" src={logo} width="70" height="70" className="d-flex align-top" />
+          <img alt="Logo" src={planet} width="70" height="70" className="d-flex align-top" />
           Space Travelers&apos; Hub
         </Navbar.Brand>
 
@@ -20,21 +20,21 @@ const Navbarmenu = () => (
           <NavLink
             to="/"
             style={({ isActive, isPending }) => ({
-              textDecoration: isActive ? 'underline' : '', paddingRight: isActive ? '1rem' : '', fontSize: isActive ? '2rem' : '1.5rem', color: isPending ? '#00b7ff' : '',
+              textDecoration: isActive ? 'underline' : '', paddingRight: isActive ? '1rem' : '', fontSize: isActive ? '1.5rem' : '1rem', color: isPending ? '#00b7ff' : '',
             })}>
             Rockets
           </NavLink>
           <NavLink
             to="/missions"
             style={({ isActive, isPending }) => ({
-              textDecoration: isActive ? 'underline' : '', paddingRight: isActive ? '1rem' : '', fontSize: isActive ? '2rem' : '1.5rem', color: isPending ? '#00b7ff' : '',
+              textDecoration: isActive ? 'underline' : '', paddingRight: isActive ? '1rem' : '', fontSize: isActive ? '1.5rem' : '1rem', color: isPending ? '#00b7ff' : '',
             })}>
             Missions
           </NavLink>
           <NavLink
             to="/profile"
             style={({ isActive, isPending }) => ({
-              textDecoration: isActive ? 'underline' : '', paddingRight: isActive ? '1rem' : '', fontSize: isActive ? '2rem' : '1.5rem', color: isPending ? '#00b7ff' : '',
+              textDecoration: isActive ? 'underline' : '', paddingRight: isActive ? '1rem' : '', fontSize: isActive ? '1.5rem' : '1rem', color: isPending ? '#00b7ff' : '',
             })}>
             My Profile
           </NavLink>
@@ -42,6 +42,7 @@ const Navbarmenu = () => (
 
       </Container>
     </Navbar>
+    <hr />
   </>
 );
