+++ src/components/Rockets.jsx
@@ -1,40 +1,8 @@
-import React, { useEffect } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
-import RocketsData from '../redux/rockets/api';
+import React from 'react';
+import RocketsList from './RocketsList';
 
-const Rockets = () => {
-  const { rockets, status, error } = useSelector((state) => state.rockets);
-  const Dispatch = useDispatch();
-  useEffect(() => {
-    Dispatch(RocketsData());
-  }, [Dispatch]);
-  if (status) {
-    return true;
-  }
-  if (error) {
-    return 'there is an error';
-  }
-  return (
-    <div className="topContainer">
-      {
-            rockets.map((rocket) => (
-              <section className="Space" key={rocket.id}>
-                <div className="img">
-                  <img src={rocket.rocket_flickr_images} alt="" />
-                </div>
-                <div className="details">
-                  <h2>{rocket.rocket_name}</h2>
-                  <p>
-                    <span />
-                    {rocket.description}
-                  </p>
-                  <button className="button" type="button">Reserve</button>
-                </div>
-              </section>
-            ))
-        }
-    </div>
-  );
-};
+const Rockets = () => (
+  <RocketsList />
+);
 
 export default Rockets;
