+++ src/styles/Rocket.css
@@ -1,3 +1,59 @@
-.img {
+.main {
+  display: flex;
+  padding-top: 20px;
+  flex-direction: column;
+  gap: 15px;
+  margin: 5px 0;
+}
+
+.card {
+  display: flex;
+  flex-direction: row;
+  font-family: 'Lato' sans-serif;
+  gap: 10px;
+  margin: 10px;
+}
+
+.myImg {
   flex: 1;
+}
+
+.rocketDetails {
+  flex: 4;
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+  margin-top: 5px;
+}
+
+.rocketDetails p {
+  padding-right: 6rem;
+  font-size: 20px;
+  line-height: 28px;
+}
+
+.myImg img {
+  width: 100%;
+}
+
+.name {
+  display: flex;
+  justify-content: flex-start;
+}
+
+button {
+  width: 15vw;
+  height: 4.5vh;
+  border: 0.2px solid black;
+  color: #fff;
+  border-radius: 5px;
+  font-size: 16px;
+  transition: 0.5s; 
+  background-color: rgb(74, 93, 255);
+}
+
+button:hover {
+  color: #fff;
+  background-color: rgb(28, 31, 56);
+  cursor: pointer;
 }
